innoweb_backend/requirements.txt
fastapi==0.118.0
uvicorn==0.32.1
prisma==0.15.0
python-dotenv==1.1.1
google-generativeai==0.8.5
requests==2.32.5
apscheduler==3.10.4
python-telegram-bot==21.6
